backend/requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
django-filter>=23.2
django-cors-headers>=4.2.0
Pillow>=10.0.0  # 用于处理图片
pandas>=2.0.0
openpyxl>=3.1.0